requirements.txt
# Core Gradio UI
gradio>=4.0.0

# LangChain and Anthropic
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-community>=0.0.1
langchain-community>=0.0.1
langchain-huggingface>=0.0.1
anthropic>=0.25.0

# Vector Database
faiss-cpu>=1.7.0

# Embeddings and ML
sentence-transformers>=2.0.0
huggingface-hub>=0.20.0
transformers>=4.30.0
torch>=2.0.0

# Google Drive Integration
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0

# Environment and Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0